[spirv] Refactoring: move classes to their own files
